+++ CHANGELOG.md
@@ -41,3 +41,7 @@ main.py
 
 0.0.8
 - added dockerfile for google cloud deployment
+
+0.0.9
+- modified logger to print to terminal allowing Google Cloud Logging to capture terminal printouts
+- created deploy.sh bash to re-upload and re-deploy docker image to cloud if changes to codebase